Added some documentation in Javadoc format

+++ Team2RecipeFinder/src/ca/ualberta/team2recipefinder/AddEditIngredientActivity.java
@@ -1,3 +1,10 @@
+/* AddEditIngredientActivity
+ * 
+ * Last Edited: March 7, 2013
+ * 
+ * 
+ */
+
 package ca.ualberta.team2recipefinder;
 
 import android.app.Activity;
@@ -9,6 +16,13 @@ import android.widget.Button;
 import android.widget.TextView;
 import android.widget.Toast;
 
+/**
+ * AddEditIngredientActivity is an android activity for adding or
+ * editing an an ingredient
+ * 
+ * @author cmput-301 team 2
+ * @see ca.ualberta.team2recipefinder.Ingredient
+ */
 public class AddEditIngredientActivity extends Activity {
 	
 	private final int EDIT = 0;
@@ -23,6 +37,11 @@ public class AddEditIngredientActivity extends Activity {
 	Button btnDelete;
 		
 	@Override
+	/**
+	 * Sets up all listeners for this activity.
+	 * 
+	 * @param	savedInstanceState Bundle containing the activity's previously frozen state, if there was one. 
+	 */
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
 		setContentView(R.layout.activity_add_edit_ingredient);
@@ -78,7 +97,7 @@ public class AddEditIngredientActivity extends Activity {
 				if (txtUnit.getText().toString().length() > 0 &&
 					txtAmount.getText().toString().length() > 0 &&
 					txtType.getText().toString().length() > 0 &&
-					Integer.parseInt(txtAmount.getText().toString()) > 0) {
+					Double.parseDouble(txtAmount.getText().toString()) > 0) {
 					
 					Ingredient ingredient = new Ingredient(txtType.getText().toString(),
 							Double.parseDouble(txtAmount.getText().toString()), txtUnit.getText().toString());
